+++ src/index.js
@@ -1,0 +1,2 @@
+// Hello there, if ur reading this, feel free to contribute! (Please, this package far from complete or bug free)
+export { default as Player } from "./classes/Player.js";
